Move helper functions to proper module
