--- lib/sanbase_web/graphql/resolvers/resolver_helpers.ex
@@ -1,13 +1,0 @@
-defmodule SanbaseWeb.Graphql.Resolvers.Helpers do
-  def error_details(changeset) do
-    changeset
-    |> Ecto.Changeset.traverse_errors(&format_error/1)
-  end
-
-  @spec format_error(Ecto.Changeset.error()) :: String.t()
-  defp format_error({msg, opts}) do
-    Enum.reduce(opts, msg, fn {key, value}, acc ->
-      String.replace(acc, "%{#{key}}", to_string(inspect(value)))
-    end)
-  end
-end
